% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsCharacterizingUnderrep.R
\name{plot.characterizing_underrep}
\alias{plot.characterizing_underrep}
\title{Plot Under represented Population Characterization}
\usage{
\method{plot}{characterizing_underrep}(x, ...)
}
\arguments{
\item{x}{A \code{characterizing_underrep} S3 object with \code{x$root$f} present as an
\code{rpart} object for the summary or characterization tree.}

\item{...}{Additional arguments passed to \code{rpart.plot::prp()}.}
}
\value{
\code{NULL}. The plot is drawn to the active graphics device.
}
\description{
Visualizes the decision tree derived from the \code{ROOT} analysis. Highlights
which subgroups are represented where \code{w = 1} versus underrepresented
where \code{w = 0} in generalization mode, or simply \code{w(x)} in \code{\{0,1\}}
in general optimization mode.
}
\examples{
\dontrun{
char.output = characterizing_underrep(diabetes_data,generalizability_path = TRUE, seed = 123)
plot(char.output)
}
}
